# tools/makeqstrdefs.py: Run qstr preprocessing in parallel.