This gives a substantial speedup of the preprocessing step (i.e. qstr.i.last).
For example on a clean build, making qstr.i.last:

21s -> 4s on STM32 (WB55)
8.9 -> 1.8s on Unix (dev).

Signed-off-by: Jim Mussared <jim.mussared@gmail.com>