@stinos 2020-10-29
Thus is probably worth it, even though the code becomes way harder to read. Why is the tempfile approach needed?