@dpgeorge 2020-10-29
Fantastic!  This is not something that could have been expressed in a Makefile...

But it seems to be failing on a lot of ports, maybe it misses processing some files?