@jimmo 2020-10-30
> But it seems to be failing on a lot of ports, maybe it misses processing some files?

True story. Just pushed a fixed.