@jimmo 2020-10-30
Final issue was python2 compatibility. Fixed.