@jimmo 2020-10-30
Updated to use an alternative implementation by @stinos using `multiprocessing.dummy.Pool` which is much simpler and easier to read.